Allow width to be specified as an argument
This commit allows the width of the image required to be passed as an
argument to the prototype2pdf binary, which then passes it to webkit2png
which is used internally.

The syntax used here is a shorthand way to specify a default without having
to use conditional statements elsewhere. This makes the syntax for using
this script as such, assuming a 150px width:

  ./prototype2pdf < urls.txt > my-prototype.pdf 150